create fn for handling unix timestamps
I created a function for handling unix timestamps. I should go through
and apply this function everywhere it will be used or any SystemTimes
are returned to the frontend.

I made the function fairly flexible by allowing users to set the
formatting that they would like.

I may want to go back and change the func so that there is a default
formatting available. I'll prob set the default formatting to be MLA
style like for what I have added here for the blogs. MLA should probably
be the style we go with if it's frontend user exposed.